Move projects and tasks controllers into api/v1 namespace
